Add start of janky testing.
We have to do something odd because we need two extra crates in order to exercise the library,
as it has to not be a proc-macro library (so it can export the functions and datastructures etc),
but many of the functions use TokenStream stuff, which only works in proc-macro context.
So we make a crate to use them in some test proc macros, and then we will make another crate to
use those proc macros in some tests.